+++ src/components/PeopleTable.tsx
@@ -1,0 +1,63 @@
+import React from 'react';
+import { useParams } from 'react-router-dom';
+import { Person } from '../types/Person';
+import { PersonLink } from './PersonLink';
+
+interface PeopleTableProps {
+  people: Person[];
+}
+
+export const PeopleTable: React.FC<PeopleTableProps> = ({ people }) => {
+  const { slug: selectedSlug } = useParams<{ slug: string }>();
+
+  const findPerson = (name: string | null): Person | null =>
+    name ? (people.find(p => p.name === name) ?? null) : null;
+
+  return (
+    <table
+      data-cy="peopleTable"
+      className="table is-striped is-hoverable is-narrow is-fullwidth"
+    >
+      <thead>
+        <tr>
+          <th>Name</th>
+          <th>Sex</th>
+          <th>Born</th>
+          <th>Died</th>
+          <th>Mother</th>
+          <th>Father</th>
+        </tr>
+      </thead>
+      <tbody>
+        {people.map(person => (
+          <tr
+            data-cy="person"
+            key={person.slug}
+            className={
+              person.slug === selectedSlug ? 'has-background-warning' : ''
+            }
+          >
+            <td>
+              <PersonLink person={person} />
+            </td>
+            <td>{person.sex}</td>
+            <td>{person.born}</td>
+            <td>{person.died}</td>
+            <td>
+              <PersonLink
+                person={findPerson(person.motherName)}
+                name={person.motherName}
+              />
+            </td>
+            <td>
+              <PersonLink
+                person={findPerson(person.fatherName)}
+                name={person.fatherName}
+              />
+            </td>
+          </tr>
+        ))}
+      </tbody>
+    </table>
+  );
+};
